fetch 분리
