+++ src/App.tsx
@@ -1,33 +1,22 @@
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 import ChangeInputEur from "./components/ChangeInputEur";
 import EurInfo from "./components/EurInfo";
+import useFetch from "./hooks/useFetch";
 
 export const App = () => {
-  const [isReady, setReady] = useState(false);
-  const [eurInfo, setEurInfo] = useState<any>();
+  const [isReady, setReady] = useState<boolean>(false);
 
-  const getEurInfo = async () => {
-    const krweur = await fetch(
-      "https://quotation-api-cdn.dunamu.com/v1/forex/recent?codes=FRX.KRWEUR"
-    )
-      .then((response) => response.json())
-      .then((array) => array[0]);
-
-    setEurInfo(krweur);
-    setReady(true);
-  };
-
-  useEffect(() => {
-    getEurInfo();
-    // return () => {};
-  }, []);
+  const eurInfos = useFetch(
+    "https://quotation-api-cdn.dunamu.com/v1/forex/recent?codes=FRX.KRWEUR",
+    setReady
+  );
 
   return (
     <>
       {isReady ? (
         <div>
-          <EurInfo eurInfo={eurInfo} />
-          <ChangeInputEur basePrice={eurInfo.basePrice} />
+          <EurInfo eurInfo={eurInfos} />
+          <ChangeInputEur basePrice={eurInfos.basePrice} />
         </div>
       ) : (
         <div>잠시만 기다려 주세요.</div>
